Improve reliability in the face of fork/exec
DxgKrnl will return ENOTTY if we have an open fd that wasn't opened
by our process. If we have a "valid" (not -1) file descriptor, but
we can't use it because of EBADF (already closed, e.g. from O_CLOEXEC)
or ENOTTY (not closed, but not from our process), then go ahead and
throw away our existing Fd and try to re-initialize it.

Note that other handles won't be able to carry forward. Attempting to
use any objects that are still alive from a fork will fail. This just
allows the newly-forked process to be able to use graphics even if
the forked-from process already was using it.